Move helpers in respective directories. Correct trace problems in testHyperRectDomain and testSimpleBoard.
