+++ examples/topology/homotopicThinning3D.cpp
@@ -35,8 +35,8 @@
 #include "DGtal/io/viewers/Viewer3D.h"
 #include "DGtal/io/DrawWithDisplay3DModifier.h"
 #include "DGtal/io/Color.h"
-#include "DGtal/helpers/Shapes.h"
 #include "DGtal/helpers/StdDefs.h"
+#include "DGtal/shapes/Shapes.h"
 
 ///////////////////////////////////////////////////////////////////////////////
