+++ src/DGtal/geometry/helpers/ContourHelper.ih
