+++ src/DGtal/topology/helpers/Surfaces.h
@@ -599,7 +599,7 @@ namespace DGtal
 
 ///////////////////////////////////////////////////////////////////////////////
 // Includes inline functions.
-#include "DGtal/helpers/Surfaces.ih"
+#include "DGtal/topology/helpers/Surfaces.ih"
 
 //                                                                           //
 ///////////////////////////////////////////////////////////////////////////////
