+++ src/DGtal/topology/helpers/Surfaces.ih
