+++ tests/io/viewers/volIVViewer.cpp
@@ -44,7 +44,7 @@
 #include "DGtal/topology/KhalimskySpaceND.h"
 #include "DGtal/helpers/StdDefs.h"
 #include "DGtal/shapes/Shapes.h"
-#include "DGtal/helpers/Surfaces.h"
+#include "DGtal/topology/helpers/Surfaces.h"
 #include "ConfigTest.h"
 
 ///////////////////////////////////////////////////////////////////////////////
