+++ tests/topology/testCellularGridSpaceND.cpp
@@ -37,7 +37,7 @@
 #include "DGtal/topology/SurfelAdjacency.h"
 #include "DGtal/topology/SurfelNeighborhood.h"
 #include "DGtal/shapes/Shapes.h"
-#include "DGtal/helpers/Surfaces.h"
+#include "DGtal/topology/helpers/Surfaces.h"
 #include "DGtal/io/Color.h"
 ///////////////////////////////////////////////////////////////////////////////
