+++ tools/2dContourTools/curvatureBC.cpp
@@ -36,8 +36,7 @@
 #include "DGtal/shapes/ShapeFactory.h"
 #include "DGtal/shapes/Shapes.h"
 #include "DGtal/helpers/StdDefs.h"
-#include "DGtal/helpers/ContourHelper.h"
-#include "DGtal/helpers/Surfaces.h"
+#include "DGtal/topology/helpers/Surfaces.h"
 
 //Grid curve
 #include "DGtal/geometry/2d/FreemanChain.h"
