+++ tools/estimators/estimatorComparator.cpp
@@ -48,7 +48,7 @@
 #include "DGtal/shapes/ShapeFactory.h"
 #include "DGtal/shapes/Shapes.h"
 #include "DGtal/helpers/StdDefs.h"
-#include "DGtal/helpers/Surfaces.h"
+#include "DGtal/topology/helpers/Surfaces.h"
 
 //Digitizer
 #include "DGtal/geometry/nd/GaussDigitizer.h"
