+++ tools/shapeGenerator/shapeGenerator.cpp
@@ -40,7 +40,7 @@
 #include "DGtal/shapes/ShapeFactory.h"
 #include "DGtal/shapes/Shapes.h"
 #include "DGtal/helpers/StdDefs.h"
-#include "DGtal/helpers/Surfaces.h"
+#include "DGtal/topology/helpers/Surfaces.h"
 
 
 #include "DGtal/io/colormaps/GrayScaleColorMap.h"
